Some inline asm correctness

+++ kernel/root.zig
@@ -38,11 +38,11 @@ pub const Ring = enum(u2) { kernel = 0, user = 3 };
 pub const PAGE_SIZE: usize = 4096;
 pub const KERNEL_BASE: usize = 0xC0000000;
 
-pub const KERNEL_CS: u8 = 1 << 3; // 0x8
-pub const KERNEL_DS: u8 = 2 << 3; // 0x10
-pub const USER_CS: u8 = (3 << 3) | 3; // 0x1B
-pub const USER_DS: u8 = (4 << 3) | 3; // 0x23
-pub const TSS: u8 = (5 << 3) | 3;
+pub const KERNEL_CS: u16 = 1 << 3; // 0x8
+pub const KERNEL_DS: u16 = 2 << 3; // 0x10
+pub const USER_CS: u16 = (3 << 3) | 3; // 0x1B
+pub const USER_DS: u16 = (4 << 3) | 3; // 0x23
+pub const TSS: u16 = (5 << 3) | 3;
 
 pub extern const kernel_end: usize;
